+++ apis/appsync/v1beta1/zz_graphqlapi_types.go
@@ -102,7 +102,7 @@ type GraphQLAPIInitParameters struct {
 type GraphQLAPILambdaAuthorizerConfigInitParameters struct {
 
 	// Number of seconds a response should be cached for. The default is 5 minutes (300 seconds). The Lambda function can override this by returning a ttlOverride key in its response. A value of 0 disables caching of responses. Minimum value of 0. Maximum value of 3600.
-	AuthorizerResultTTLInSeconds *int64 `json:"authorizerResultTtlInSeconds,omitempty" tf:"authorizer_result_ttl_in_seconds,omitempty"`
+	AuthorizerResultTTLInSeconds *float64 `json:"authorizerResultTtlInSeconds,omitempty" tf:"authorizer_result_ttl_in_seconds,omitempty"`
 
 	// ARN of the Lambda function to be called for authorization. Note: This Lambda function must have a resource-based policy assigned to it, to allow lambda:InvokeFunction from service principal appsync.amazonaws.com.
 	AuthorizerURI *string `json:"authorizerUri,omitempty" tf:"authorizer_uri,omitempty"`
@@ -114,7 +114,7 @@ type GraphQLAPILambdaAuthorizerConfigInitParameters struct {
 type GraphQLAPILambdaAuthorizerConfigObservation struct {
 
 	// Number of seconds a response should be cached for. The default is 5 minutes (300 seconds). The Lambda function can override this by returning a ttlOverride key in its response. A value of 0 disables caching of responses. Minimum value of 0. Maximum value of 3600.
-	AuthorizerResultTTLInSeconds *int64 `json:"authorizerResultTtlInSeconds,omitempty" tf:"authorizer_result_ttl_in_seconds,omitempty"`
+	AuthorizerResultTTLInSeconds *float64 `json:"authorizerResultTtlInSeconds,omitempty" tf:"authorizer_result_ttl_in_seconds,omitempty"`
 
 	// ARN of the Lambda function to be called for authorization. Note: This Lambda function must have a resource-based policy assigned to it, to allow lambda:InvokeFunction from service principal appsync.amazonaws.com.
 	AuthorizerURI *string `json:"authorizerUri,omitempty" tf:"authorizer_uri,omitempty"`
@@ -127,7 +127,7 @@ type GraphQLAPILambdaAuthorizerConfigParameters struct {
 
 	// Number of seconds a response should be cached for. The default is 5 minutes (300 seconds). The Lambda function can override this by returning a ttlOverride key in its response. A value of 0 disables caching of responses. Minimum value of 0. Maximum value of 3600.
 	// +kubebuilder:validation:Optional
-	AuthorizerResultTTLInSeconds *int64 `json:"authorizerResultTtlInSeconds,omitempty" tf:"authorizer_result_ttl_in_seconds,omitempty"`
+	AuthorizerResultTTLInSeconds *float64 `json:"authorizerResultTtlInSeconds,omitempty" tf:"authorizer_result_ttl_in_seconds,omitempty"`
 
 	// ARN of the Lambda function to be called for authorization. Note: This Lambda function must have a resource-based policy assigned to it, to allow lambda:InvokeFunction from service principal appsync.amazonaws.com.
 	// +kubebuilder:validation:Optional
@@ -186,13 +186,13 @@ type GraphQLAPIObservation struct {
 type GraphQLAPIOpenIDConnectConfigInitParameters struct {
 
 	// Number of milliseconds a token is valid after being authenticated.
-	AuthTTL *int64 `json:"authTtl,omitempty" tf:"auth_ttl,omitempty"`
+	AuthTTL *float64 `json:"authTtl,omitempty" tf:"auth_ttl,omitempty"`
 
 	// Client identifier of the Relying party at the OpenID identity provider. This identifier is typically obtained when the Relying party is registered with the OpenID identity provider. You can specify a regular expression so the AWS AppSync can validate against multiple client identifiers at a time.
 	ClientID *string `json:"clientId,omitempty" tf:"client_id,omitempty"`
 
 	// Number of milliseconds a token is valid after being issued to a user.
-	IatTTL *int64 `json:"iatTtl,omitempty" tf:"iat_ttl,omitempty"`
+	IatTTL *float64 `json:"iatTtl,omitempty" tf:"iat_ttl,omitempty"`
 
 	// Issuer for the OpenID Connect configuration. The issuer returned by discovery MUST exactly match the value of iss in the ID Token.
 	Issuer *string `json:"issuer,omitempty" tf:"issuer,omitempty"`
@@ -201,13 +201,13 @@ type GraphQLAPIOpenIDConnectConfigInitParameters struct {
 type GraphQLAPIOpenIDConnectConfigObservation struct {
 
 	// Number of milliseconds a token is valid after being authenticated.
-	AuthTTL *int64 `json:"authTtl,omitempty" tf:"auth_ttl,omitempty"`
+	AuthTTL *float64 `json:"authTtl,omitempty" tf:"auth_ttl,omitempty"`
 
 	// Client identifier of the Relying party at the OpenID identity provider. This identifier is typically obtained when the Relying party is registered with the OpenID identity provider. You can specify a regular expression so the AWS AppSync can validate against multiple client identifiers at a time.
 	ClientID *string `json:"clientId,omitempty" tf:"client_id,omitempty"`
 
 	// Number of milliseconds a token is valid after being issued to a user.
-	IatTTL *int64 `json:"iatTtl,omitempty" tf:"iat_ttl,omitempty"`
+	IatTTL *float64 `json:"iatTtl,omitempty" tf:"iat_ttl,omitempty"`
 
 	// Issuer for the OpenID Connect configuration. The issuer returned by discovery MUST exactly match the value of iss in the ID Token.
 	Issuer *string `json:"issuer,omitempty" tf:"issuer,omitempty"`
@@ -217,7 +217,7 @@ type GraphQLAPIOpenIDConnectConfigParameters struct {
 
 	// Number of milliseconds a token is valid after being authenticated.
 	// +kubebuilder:validation:Optional
-	AuthTTL *int64 `json:"authTtl,omitempty" tf:"auth_ttl,omitempty"`
+	AuthTTL *float64 `json:"authTtl,omitempty" tf:"auth_ttl,omitempty"`
 
 	// Client identifier of the Relying party at the OpenID identity provider. This identifier is typically obtained when the Relying party is registered with the OpenID identity provider. You can specify a regular expression so the AWS AppSync can validate against multiple client identifiers at a time.
 	// +kubebuilder:validation:Optional
@@ -225,7 +225,7 @@ type GraphQLAPIOpenIDConnectConfigParameters struct {
 
 	// Number of milliseconds a token is valid after being issued to a user.
 	// +kubebuilder:validation:Optional
-	IatTTL *int64 `json:"iatTtl,omitempty" tf:"iat_ttl,omitempty"`
+	IatTTL *float64 `json:"iatTtl,omitempty" tf:"iat_ttl,omitempty"`
 
 	// Issuer for the OpenID Connect configuration. The issuer returned by discovery MUST exactly match the value of iss in the ID Token.
 	// +kubebuilder:validation:Optional
@@ -339,7 +339,7 @@ type GraphQLAPIUserPoolConfigParameters struct {
 type LambdaAuthorizerConfigInitParameters struct {
 
 	// Number of seconds a response should be cached for. The default is 5 minutes (300 seconds). The Lambda function can override this by returning a ttlOverride key in its response. A value of 0 disables caching of responses. Minimum value of 0. Maximum value of 3600.
-	AuthorizerResultTTLInSeconds *int64 `json:"authorizerResultTtlInSeconds,omitempty" tf:"authorizer_result_ttl_in_seconds,omitempty"`
+	AuthorizerResultTTLInSeconds *float64 `json:"authorizerResultTtlInSeconds,omitempty" tf:"authorizer_result_ttl_in_seconds,omitempty"`
 
 	// ARN of the Lambda function to be called for authorization. Note: This Lambda function must have a resource-based policy assigned to it, to allow lambda:InvokeFunction from service principal appsync.amazonaws.com.
 	AuthorizerURI *string `json:"authorizerUri,omitempty" tf:"authorizer_uri,omitempty"`
@@ -351,7 +351,7 @@ type LambdaAuthorizerConfigInitParameters struct {
 type LambdaAuthorizerConfigObservation struct {
 
 	// Number of seconds a response should be cached for. The default is 5 minutes (300 seconds). The Lambda function can override this by returning a ttlOverride key in its response. A value of 0 disables caching of responses. Minimum value of 0. Maximum value of 3600.
-	AuthorizerResultTTLInSeconds *int64 `json:"authorizerResultTtlInSeconds,omitempty" tf:"authorizer_result_ttl_in_seconds,omitempty"`
+	AuthorizerResultTTLInSeconds *float64 `json:"authorizerResultTtlInSeconds,omitempty" tf:"authorizer_result_ttl_in_seconds,omitempty"`
 
 	// ARN of the Lambda function to be called for authorization. Note: This Lambda function must have a resource-based policy assigned to it, to allow lambda:InvokeFunction from service principal appsync.amazonaws.com.
 	AuthorizerURI *string `json:"authorizerUri,omitempty" tf:"authorizer_uri,omitempty"`
@@ -364,7 +364,7 @@ type LambdaAuthorizerConfigParameters struct {
 
 	// Number of seconds a response should be cached for. The default is 5 minutes (300 seconds). The Lambda function can override this by returning a ttlOverride key in its response. A value of 0 disables caching of responses. Minimum value of 0. Maximum value of 3600.
 	// +kubebuilder:validation:Optional
-	AuthorizerResultTTLInSeconds *int64 `json:"authorizerResultTtlInSeconds,omitempty" tf:"authorizer_result_ttl_in_seconds,omitempty"`
+	AuthorizerResultTTLInSeconds *float64 `json:"authorizerResultTtlInSeconds,omitempty" tf:"authorizer_result_ttl_in_seconds,omitempty"`
 
 	// ARN of the Lambda function to be called for authorization. Note: This Lambda function must have a resource-based policy assigned to it, to allow lambda:InvokeFunction from service principal appsync.amazonaws.com.
 	// +kubebuilder:validation:Optional
@@ -424,13 +424,13 @@ type LogConfigParameters struct {
 type OpenIDConnectConfigInitParameters struct {
 
 	// Number of milliseconds a token is valid after being authenticated.
-	AuthTTL *int64 `json:"authTtl,omitempty" tf:"auth_ttl,omitempty"`
+	AuthTTL *float64 `json:"authTtl,omitempty" tf:"auth_ttl,omitempty"`
 
 	// Client identifier of the Relying party at the OpenID identity provider. This identifier is typically obtained when the Relying party is registered with the OpenID identity provider. You can specify a regular expression so the AWS AppSync can validate against multiple client identifiers at a time.
 	ClientID *string `json:"clientId,omitempty" tf:"client_id,omitempty"`
 
 	// Number of milliseconds a token is valid after being issued to a user.
-	IatTTL *int64 `json:"iatTtl,omitempty" tf:"iat_ttl,omitempty"`
+	IatTTL *float64 `json:"iatTtl,omitempty" tf:"iat_ttl,omitempty"`
 
 	// Issuer for the OpenID Connect configuration. The issuer returned by discovery MUST exactly match the value of iss in the ID Token.
 	Issuer *string `json:"issuer,omitempty" tf:"issuer,omitempty"`
@@ -439,13 +439,13 @@ type OpenIDConnectConfigInitParameters struct {
 type OpenIDConnectConfigObservation struct {
 
 	// Number of milliseconds a token is valid after being authenticated.
-	AuthTTL *int64 `json:"authTtl,omitempty" tf:"auth_ttl,omitempty"`
+	AuthTTL *float64 `json:"authTtl,omitempty" tf:"auth_ttl,omitempty"`
 
 	// Client identifier of the Relying party at the OpenID identity provider. This identifier is typically obtained when the Relying party is registered with the OpenID identity provider. You can specify a regular expression so the AWS AppSync can validate against multiple client identifiers at a time.
 	ClientID *string `json:"clientId,omitempty" tf:"client_id,omitempty"`
 
 	// Number of milliseconds a token is valid after being issued to a user.
-	IatTTL *int64 `json:"iatTtl,omitempty" tf:"iat_ttl,omitempty"`
+	IatTTL *float64 `json:"iatTtl,omitempty" tf:"iat_ttl,omitempty"`
 
 	// Issuer for the OpenID Connect configuration. The issuer returned by discovery MUST exactly match the value of iss in the ID Token.
 	Issuer *string `json:"issuer,omitempty" tf:"issuer,omitempty"`
@@ -455,7 +455,7 @@ type OpenIDConnectConfigParameters struct {
 
 	// Number of milliseconds a token is valid after being authenticated.
 	// +kubebuilder:validation:Optional
-	AuthTTL *int64 `json:"authTtl,omitempty" tf:"auth_ttl,omitempty"`
+	AuthTTL *float64 `json:"authTtl,omitempty" tf:"auth_ttl,omitempty"`
 
 	// Client identifier of the Relying party at the OpenID identity provider. This identifier is typically obtained when the Relying party is registered with the OpenID identity provider. You can specify a regular expression so the AWS AppSync can validate against multiple client identifiers at a time.
 	// +kubebuilder:validation:Optional
@@ -463,7 +463,7 @@ type OpenIDConnectConfigParameters struct {
 
 	// Number of milliseconds a token is valid after being issued to a user.
 	// +kubebuilder:validation:Optional
-	IatTTL *int64 `json:"iatTtl,omitempty" tf:"iat_ttl,omitempty"`
+	IatTTL *float64 `json:"iatTtl,omitempty" tf:"iat_ttl,omitempty"`
 
 	// Issuer for the OpenID Connect configuration. The issuer returned by discovery MUST exactly match the value of iss in the ID Token.
 	// +kubebuilder:validation:Optional
