+++ apis/backup/v1beta1/zz_reportplan_types.go
@@ -144,7 +144,7 @@ type ReportSettingInitParameters struct {
 	FrameworkArns []*string `json:"frameworkArns,omitempty" tf:"framework_arns,omitempty"`
 
 	// Specifies the number of frameworks a report covers.
-	NumberOfFrameworks *int64 `json:"numberOfFrameworks,omitempty" tf:"number_of_frameworks,omitempty"`
+	NumberOfFrameworks *float64 `json:"numberOfFrameworks,omitempty" tf:"number_of_frameworks,omitempty"`
 
 	// Specifies the list of Organizational Units a report covers.
 	OrganizationUnits []*string `json:"organizationUnits,omitempty" tf:"organization_units,omitempty"`
@@ -165,7 +165,7 @@ type ReportSettingObservation struct {
 	FrameworkArns []*string `json:"frameworkArns,omitempty" tf:"framework_arns,omitempty"`
 
 	// Specifies the number of frameworks a report covers.
-	NumberOfFrameworks *int64 `json:"numberOfFrameworks,omitempty" tf:"number_of_frameworks,omitempty"`
+	NumberOfFrameworks *float64 `json:"numberOfFrameworks,omitempty" tf:"number_of_frameworks,omitempty"`
 
 	// Specifies the list of Organizational Units a report covers.
 	OrganizationUnits []*string `json:"organizationUnits,omitempty" tf:"organization_units,omitempty"`
@@ -189,7 +189,7 @@ type ReportSettingParameters struct {
 
 	// Specifies the number of frameworks a report covers.
 	// +kubebuilder:validation:Optional
-	NumberOfFrameworks *int64 `json:"numberOfFrameworks,omitempty" tf:"number_of_frameworks,omitempty"`
+	NumberOfFrameworks *float64 `json:"numberOfFrameworks,omitempty" tf:"number_of_frameworks,omitempty"`
 
 	// Specifies the list of Organizational Units a report covers.
 	// +kubebuilder:validation:Optional
